More error checking

+++ scripts/quick_start.sh
@@ -412,19 +412,24 @@ function demos(){
 
 # Reset the project back to a clean slate
 function reset_local(){
-    printf "\e[0;34m\nStop your previous docker-compose project? \e[0m"
-    read STOP_COMPOSE
 
-    # If true, stop any docker-compose projects built with this project.
-    case $STOP_COMPOSE in
-    y|Y|yes)
-        cd ${PROJECT_ROOT}
-        docker-compose down
-        cd - >/dev/null 2>&1
-    ;;
-    n|N|no)
-    ;;
-    esac
+    # If compose already has a vault running, ask if we should stop
+    if docker-compose ps | grep -q vault > /dev/null 2>&1;
+    then
+        printf "\e[0;34m\nStop your previous docker-compose project? \e[0m"
+        read STOP_COMPOSE
+
+        # If true, stop any docker-compose projects built with this project.
+        case $STOP_COMPOSE in
+        y|Y|yes)
+            cd ${PROJECT_ROOT}
+            docker-compose down
+            cd - >/dev/null 2>&1
+        ;;
+        n|N|no)
+        ;;
+        esac
+    fi
 
     printf "\e[0;34m\nShould all previously used files be removed?\e[0m i.e Vault and Consul data, Terraform backends, and TLS Certs? "
     read RESET_BOOL
@@ -503,7 +508,7 @@ export VAULT_SKIP_VERIFY=true
 VAULT_ADDRESS="https://127.0.0.1:8200"
 export VAULT_ADDR=${VAULT_ADDRESS}
 
-# Absolute path to the project root
+# Absolute path to the project root based on this scripts location. 
 PROJECT_ROOT=$(dirname $(cd `dirname $0` && pwd))
 # File to store the restore/recovery keys and the root token
 KEYS_FILE="${PROJECT_ROOT}/_data/keys.txt"
@@ -512,7 +517,7 @@ BUCKET_NAME="ian-bucket-dev"
 # Name of the license file in the AWS bucket above
 LICENSE_FILE="license.txt"
 
-# Terraform and Vault CLI check
+# Dependency check, both OSS and ENT Vault Versions. 
 if ! which vault >/dev/null 2>&1
 then
     printf "\e[0;34m\nVault not installed, please install to continue...\e[0m\n\n"
@@ -520,6 +525,10 @@ then
 elif ! which terraform >/dev/null 2>&1
 then
     printf "\e[0;34m\nTerraform not installed, please install to continue...\e[0m\n\n"
+    exit 0
+elif ! which docker >/dev/null 2>&1
+then
+    printf "\e[0;34m\nDocker not installed, please install to continue...\e[0m\n\n"
     exit 0
 fi
 
@@ -536,6 +545,10 @@ read MAIN_MENU
 
 case ${MAIN_MENU} in
 1)
+    #######################
+    ### Full Script Run ###
+    #######################
+
     # Clean old files and compose projects
     if ls ${PROJECT_ROOT}/_data >/dev/null 2>&1;
     then
@@ -583,6 +596,10 @@ case ${MAIN_MENU} in
 
     case ${VAULT_VERSION} in
     1)
+        #################
+        ### OSS Vault ###
+        #################
+
         # Check if docker is already running with a vault image
         if ! docker ps 2>/dev/null | grep -q "vault";
         then
@@ -622,6 +639,10 @@ case ${MAIN_MENU} in
         export VAULT_TOKEN="${VR_TOKEN}"
     ;;
     2)
+        #################
+        ### Ent Vault ###
+        #################
+
         # Check if docker is already running with a vault image
         if ! docker ps 2>/dev/null | grep -q "vault"; then
 
@@ -730,6 +751,10 @@ EOF
     demos
 ;;
 2)
+    ##########################################
+    ### Bootstrap An Already Running Vault ###
+    ##########################################
+
     PROJECT_NAME=$(ls ${PROJECT_ROOT}/config/cluster_certs/ | grep -v localhost | awk -F'.' '/crt/ {print $1}')
     export TF_VAR_env=${PROJECT_NAME}
 
@@ -746,6 +771,10 @@ EOF
     fi
 ;;
 3)
+    #################
+    ### Run Demos ###
+    #################
+
     PROJECT_NAME=$(ls ${PROJECT_ROOT}/config/cluster_certs/ | grep -v localhost | awk -F'.' '/crt/ {print $1}')
     export TF_VAR_env=${PROJECT_NAME}
 
@@ -759,12 +788,23 @@ EOF
     fi
 ;;
 4)
+    #####################
+    ## Reset Local Env ##
+    #####################
+
     reset_local
 ;;
 5)
+    #######################
+    ### Exit The Script ###
+    #######################
     exit 0
 ;;
 *)
+    ######################
+    ### Error Catching ###
+    ######################
+
     printf "\e[0;34m\nInvalid Selection, please try again.\n\n"
     ${PROJECT_ROOT}/scripts/$(basename $0) && exit
 ;;
